requirements.txt
# Multi-provider LLM interface
litellm>=1.50.0
python-dotenv>=1.0.0

# Jupyter Book (static site generator)
jupyter-book>=1.0.0
ghp-import>=2.1.0

# Notebooks
jupyter>=1.0.0
ipykernel>=6.29.0
tiktoken>=0.7.0
numpy>=1.26.0
pydantic>=2.0.0
scikit-learn>=1.4.0
matplotlib>=3.8.0

# RAG (Module 4)
chromadb>=0.4.0

# Fine-tuning (Module 6) — heavy deps, install separately if needed
# pip install transformers>=4.40.0 datasets>=2.0.0 peft>=0.10.0 trl>=0.8.0 torch>=2.1.0
